AI_Powered_Legal_Chatbot/backend/requirements.txt
langchain 
langchain-community 
langchain-google-genai 
python-dotenv 
tqdm
faiss-cpu
groq
langchain-groq
PyPDF2
streamlit
pypdf
fastapi
uvicorn
google-generative-ai